pyproject.toml
[tool.poetry]
name = "test-messenger"
version = "0.1.0"
description = ""
authors = ["VitaliyDrozdov <vitdrzdv@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

fastapi = {extras = ["standard"], version = "^0.115.8"}
slugify = "^0.0.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.38"
pydantic-settings = "^2.8.1"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
isort = "^6.0.0"
flake8 = "^7.1.2"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
